+++ lib/presentation/screen/component/schedule_list_view.dart
@@ -7,7 +7,7 @@ import 'package:provider/provider.dart';
 import '../../../domain/entity/schedule.dart';
 import '../../const/strings.dart';
 import '../provider/schedule_provider.dart';
-import 'date_control_dialog.dart';
+import 'date_control_dialog/date_control_dialog.dart';
 import 'default_component.dart';
 
 class ScheduleListView extends StatelessWidget {
